+++ CHANGELOG.md
@@ -2,14 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.0.1-RC24]
+## [0.0.1-RC26]
 
 - Adds support for handling Generics in `SchemaBuilder` & `SchemaValidationUtils`
 - SchemaDetails: Use set for tags
 - LeiaMessageProduceClient: Append schema tags for multiplexed messages
 - TagBasedNameResolver: Bug Fix: Transform backend name to uppercase before comparison
 
+## [0.0.1-RC25]
+
+- Adds support for handling Generics in `SchemaBuilder` & `SchemaValidationUtils`
+- Exposes Schema Index via Elastic Config
+
 ## [0.0.1-RC23]
+
 - Removed leia-refresher and replaced it with [korg](github.com/grookage/korg)
 - Fixed an AS client bug for expression evaluation - [MR](https://github.com/grookage/leia/pull/35)
 - Added orgId and tenant to schemaKey and has removed random flattened out structures, but instead maintaining storageRecord separately from application model.